Feature/postgres support

+++ BugTracker/Models/BugModel.cs
@@ -1,9 +1,11 @@
 ﻿using System;
+using System.ComponentModel.DataAnnotations;
+
 namespace BugTracker.Models
 {
     public class Bug
     {
-        public int Id { get; set; }
+	    public long Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
     }
